restructured the tests
added a new structure for the backend and frontend tests. Additionally there are tests that can be run on both environments and therefore they go to the common folder.
Lastly, common\test_headers.py contains helpers functions for some tests that need to be run slightly differently in backend and frontend